+++ lib_xcore_math/src/arch/ref/chunk.c
@@ -75,7 +75,7 @@ void chunk_q30_power_series(
     int32_t acc = 0;
 
     int32_t pow = 0x40000000;
-    for(size_t i = 0; i < term_count; i++){
+    for(unsigned i = 0; i < term_count; i++){
       acc = vlmacc32(acc, pow, coef[8*i]);
       pow = vlmul32(pow, b[k]);
     }
