+++ lib_xcore_math/src/arch/ref/vect_s16_extract.c
@@ -9,16 +9,12 @@
 #include "xmath/xs3/vpu_scalar_ops.h"
 
 
-
-
-
-
 void vect_s16_extract_high_byte(
     int8_t a[],
     const int16_t b[],
     const unsigned len)
 {
-  for(size_t k = 0; k < len; k++){
+  for(unsigned k = 0; k < len; k++){
     a[k] = (b[k] & 0xFF00) >> 8;
   }
 }
@@ -28,7 +24,7 @@ void vect_s16_extract_low_byte(
     const int16_t b[],
     const unsigned len)
 {
-  for(size_t k = 0; k < len; k++){
+  for(unsigned k = 0; k < len; k++){
     a[k] = (b[k] & 0x00FF);
   }
 }
