+++ lib_xcore_math/src/filter/filters.c
@@ -76,7 +76,7 @@ int32_t filter_biquads_s32(
 {
     int32_t smp = new_sample;
 
-    for(size_t i = 0; i < block_count; i++)
+    for(unsigned i = 0; i < block_count; i++)
         smp = filter_biquad_s32(&biquads[i], smp);
     
     return smp;
